uintmax_t

+++ proto.h
@@ -1284,7 +1284,7 @@ Perl_grok_numeric_radix(pTHX_ const char **sp, const char *send)
         assert(sp); assert(send)
 
 PERL_CALLCONV UV
-Perl_grok_uint_by_base(pTHX_ const char * const start, STRLEN *len_p, I32 *flags, NV *result, STRLEN *digit_count, U8 max_permissible_bits, unsigned shift1, unsigned shift2, const U8 lookup_bit, const U8 offset);
+Perl_grok_uint_by_base(pTHX_ const char * const start, STRLEN *len_p, I32 *flags, NV *result, STRLEN *digit_count, U8 max_permissible_bits, uint_fast8_t shift1, uint_fast8_t shift2, const U8 lookup_bit, const U8 offset);
 #define PERL_ARGS_ASSERT_GROK_UINT_BY_BASE      \
         assert(start); assert(len_p); assert(flags)
